requirements.txt
boto3==1.26.114
EdgeGPT==0.1.23
openai==0.27.4
openai_whisper==20230314
pydub==0.25.1
SpeechRecognition==3.10.0
PyAudio==0.2.13